chore: rename demo folder into packages/spec
