+++ package.json
@@ -4,11 +4,11 @@
   "license": "MIT",
   "private": true,
   "scripts": {
-    "start": "yarn --cwd demo hops start",
+    "start": "yarn --cwd packages/spec hops start",
     "release": "lerna publish",
     "release:next": "lerna publish prerelease --preid=rc --npm-tag=next",
-    "test": "yarn lint && yarn --cwd demo jest",
-    "lint": "eslint .",
+    "test": "yarn lint && yarn --cwd packages/spec jest",
+    "lint": "eslint ./packages",
     "fmt": "prettier --write '**/*.{js,json,css}' '**/README.md'"
   },
   "engines": {
