+++ packages/spec/content.mdx
@@ -1,6 +1,6 @@
 # Hops-MDX!
 
-![MDX Logo](./files/mdx-logo.svg)
+![MDX Logo](./mdx-logo.svg)
 
 Hello **world**! :+1:
