+++ packages/spec/package.json
@@ -24,26 +24,22 @@
   },
   "jest": {
     "projects": [
-      "<rootDir>/spec/helper",
-      "<rootDir>/spec/preset"
+      "<rootDir>/helper",
+      "<rootDir>/preset"
     ]
   },
   "dependencies": {
     "hops": "^11.0.1",
     "hops-mdx": "^0.0.1-rc.1",
+    "hops-mdx-jest": "^0.0.1-rc.1",
+    "jest": "^23.4.2",
+    "jest-preset-hops": "^11.0.1",
     "react": "^16.4.2",
     "react-dom": "^16.4.2",
     "react-helmet": "^5.2.0",
     "react-hot-loader": "^4.6.0",
     "react-router-dom": "^4.3.1",
+    "react-test-renderer": "^16.4.2",
     "remark-emoji": "^2.0.2"
-  },
-  "devDependencies": {
-    "got": "^9.5.0",
-    "hops-mdx-jest": "^0.0.1-rc.1",
-    "jest": "^23.4.2",
-    "jest-dev-server": "^3.7.0",
-    "jest-preset-hops": "^11.0.1",
-    "react-test-renderer": "^16.4.2"
   }
 }
