+++ packages/spec/preset/global-setup.js
@@ -1,6 +1,4 @@
-// eslint-disable-next-line node/no-extraneous-require
 const { setup: setupDevServer } = require('jest-dev-server');
-// eslint-disable-next-line node/no-extraneous-require
 const got = require('got');
 
 const { PORT = '8081' } = process.env;
@@ -8,7 +6,7 @@ const { PORT = '8081' } = process.env;
 module.exports = async () => {
   await setupDevServer({
     command: `PORT=${PORT} yarn start`,
-    launchTimeout: 60000,
+    launchTimeout: 100000,
     port: Number(PORT),
     usedPortAction: 'kill',
   });
